+++ tests/cpu.rs
@@ -1,5 +1,3 @@
-
-
 #[allow(unused_imports)]
 use spork::{Platform, Spork, SporkError, SporkErrorKind, StatType, Stats};
 
@@ -25,7 +23,7 @@ fn fib(n: u64) -> u64 {
 
 fn now_ms() -> i64 {
     let now = Utc::now();
-    (now.timestamp() * 1000 + (now.timestamp_subsec_millis() as i64)) as i64
+    now.timestamp() * 1000 + (now.timestamp_subsec_millis() as i64)
 }
 
 fn rand_in_range(l: u64, r: u64) -> u64 {
